refactor(web): extract managed-files routes into web_routers/files (web_server.py god-file slice R2-C1)
